% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_rasters.R
\name{categorize_rasters}
\alias{categorize_rasters}
\title{Categorize rasters}
\usage{
categorize_rasters(raster_data, data)
}
\arguments{
\item{raster_data}{Raster stack to be dummified.}

\item{data}{Modeling data. Output of \code{format_data()} function.}
}
\value{
RasterStack
}
\description{
This function creates dummy variables for categorical variable levels.
}
\examples{
None
}
