% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{clean_coeffs}
\alias{clean_coeffs}
\title{Clean up model coefficients output}
\usage{
clean_coeffs(model, r_data, v_rasters, m_assessment)
}
\arguments{
\item{model}{Model object.}

\item{r_data}{Raw raster data}

\item{v_rasters}{VIF raster values}

\item{m_assessment}{Model assessment object}
}
\value{
dataframe
}
\description{
Clean up model coefficients output
}
\examples{
clean_coeffs(model = my_model, r_data = raster_data_raw, v_rasters = vif_rasters, m_assessment = model_assessment)
}
