% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate model
Get evaluation metrics (currently AUC and TSS) for agricultural conversion models.}
\usage{
evaluate_model(fitted_model, data)
}
\arguments{
\item{fitted_model}{glm model. Output of fit_model() function (or glm() function in that matter). Any model of class "glm".}

\item{data}{- data for modeling. Output of format_data() function. List with "evaluation_data" slot.}
}
\value{
dataframe with TSS and AUC values.
}
\description{
Evaluate model
Get evaluation metrics (currently AUC and TSS) for agricultural conversion models.
}
\examples{
None.
}
