% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data}
\usage{
format_data(explanatory_variables, response_variable, evaluation_data,
  VIF_select = FALSE, cross_validate = FALSE, threshold = 10)
}
\arguments{
\item{explanatory_variables}{Raster data, output from get_rasters()}

\item{response_variable}{Data for fitting, output from load_PA()}

\item{evaluation_data}{Data for evaluation, output from load_PA(). Necessary if cross_validate = FALSE}

\item{VIF_select}{Do VIF selection of predictiors (only works now for cross_validate = TRUE). Not necessary, since VIF selection is now done a priori.}

\item{cross_validate}{Assess models on cross validated or hind casted dataset. Default is FALSE.}

\item{threshold}{Thershold for VIF calculation. Not needed anymore (see VIF_select arg).}
}
\value{
List of 2. Training data and evaluation data.
}
\description{
Prepares data for modeling. Gives list with training and evaluation data needed for the model.
}
\examples{
None.
}
