% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_absences.R
\name{get_absences}
\alias{get_absences}
\title{Get absences}
\usage{
get_absences(raster_old, raster_new, category, abs_number,
  exclude_urban = FALSE, sp = FALSE)
}
\arguments{
\item{raster_old}{Raster from the previous timestep.}

\item{raster_new}{Raster from the future timestep.}

\item{category}{Modeled land cover category. Currently only agriculture is supported. Numeric. One of c(10, 30, 40).}

\item{abs_number}{Number of absences to create. Default is 10 000.}

\item{exclude_urban}{Should the urban areas be excluded from the absences calculation. If \code{exclude_urban = TRUE},
absences will not be created in urban areas (class 190) in addition to modeled category. Default is FALSE.}

\item{sp}{Logical. If \code{sp = TRUE}, spatial object of class sp will be returned. Otherwise a dataframe (default).}
}
\value{
Dataframe or sp object (depending on the \code{sp} argument).
}
\description{
Take land cover rasters from two timesteps (old and new).
}
\examples{
None.
}
