% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change_raster_GDAL.R
\name{get_change_raster_GDAL}
\alias{get_change_raster_GDAL}
\title{Extract changed cells from two rasters
Get changes from two rasters from different periods. The function is suitable for parallel processing of very large rasters on computers with many cores.}
\usage{
get_change_raster_GDAL(x, y, size, outpath, outfile, category,
  number_of_cores = 1)
}
\arguments{
\item{x}{t1 raster}

\item{y}{t2 raster}

\item{size}{Tilesize. If blank, tile size will be \code{1000 * cell resolution}.}

\item{outpath}{Folder in which the temporary raster files will be created. Temporary files will be deleted when outfile is created.}

\item{outfile}{Output file name. Optional.}

\item{category}{Land cover category to extract.}

\item{number_of_cores}{Number of cores to run. Default is 1.}
}
\value{
None.
}
\description{
Extract changed cells from two rasters
Get changes from two rasters from different periods. The function is suitable for parallel processing of very large rasters on computers with many cores.
}
\examples{
Dunno
}
