% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{get_rasters}
\alias{get_rasters}
\title{Load raster data}
\usage{
get_rasters(region)
}
\arguments{
\item{region}{Region for which to get the rasters.}
}
\value{
RasterStack
}
\description{
This function loads raster data (predictor variables) necessary for agri modeling.
File paths are hardcoded.
}
\examples{
get_rasters("Korea_region")
}
