% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PA.R
\name{load_PA}
\alias{load_PA}
\title{Load response variable}
\usage{
load_PA(region, category, type, path)
}
\arguments{
\item{region}{Which region.}

\item{category}{Which category.}

\item{type}{Hind-casted or cross-validated type of data. One of \code{c("Fit", "Eval")}}

\item{path}{Folder path in which to search for the presence/absence data. Usually located in "./LU_data/Changes_*/".}
}
\value{
Dataframe
}
\description{
Load response variable (presences/absences) for specific region, category and assessment method.
}
\examples{
None so far
}
