% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PA.R
\name{load_PA2}
\alias{load_PA2}
\title{Load data2}
\usage{
load_PA2(region, category, type)
}
\arguments{
\item{region}{Which region.}

\item{category}{Which category.}

\item{type}{Fit or eval.}
}
\value{
Dataframe
}
\description{
Second version of the function, hard coded file paths should be removed.
}
\examples{
None.
}
