% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarify_points.R
\name{rarify_points}
\alias{rarify_points}
\title{Rarify points}
\usage{
rarify_points(point_data, raster_data, points_number = 1, return = "sf")
}
\arguments{
\item{point_data}{Point data to rarify. Should be of \emph{class sp}.}

\item{raster_data}{Raster grid used to sample points. \emph{Raster*} object.}

\item{points_number}{Number of points to sample per grid cell. Default is 1. \strong{Using points_number > 1 is still experimental.}}

\item{return}{Type of object to return. One of \code{c("sf", "sp")}. Default is \emph{"sf"}.}
}
\value{
Spatial point data. Either sp or sf class.
}
\description{
Take a set of spatial points data, and sample them according the the raster grid.
Points will be sampled so a specified number of points per grid cell remains.
Number of points is defined by \code{points_number} argument (still experimental if points_number > 1).
}
\examples{
library(LUpak)

bioclim_mask <- raster(milkunize("Temp/Korea_mask_crop.tif"))
korea_points <-  shapefile(milkunize("Temp/Korea_points_crop.shp"))

my_crops_rarified <- rarify_points(korea_points, bioclim_mask)
}
