% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_absences.R
\name{rcl_excluded_vals}
\alias{rcl_excluded_vals}
\title{Reclassify excluded values}
\usage{
rcl_excluded_vals(x, vals_included, vals_excluded, filename)
}
\arguments{
\item{x}{Input raster to be classified.}

\item{vals_included}{Values of all possible categories (without the excluded values).}

\item{vals_excluded}{Values to be excluded from absence creation. Usually modeled category or urban.}

\item{filename}{Temporary file to write.}
}
\value{
Raster object.
}
\description{
This is a helper function for absences creation that takes a raster, and reclassifies it to 0-1 raster.
Reclassified raster should be ESA CCI LC data. For the \code{values_included} argument,
numerical vector of all land cover values in the raster (they are turned to value 1).
For the \code{values_excluded} argument, numerical vector of land cover categories to be excluded
from absence creation (usually the modeled category and urban/bare areas)
should be provided (they are turned to value 0).
Argument \code{filename} is the name of the temporary file created.
}
\examples{
None.
}
