% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_absences.R
\name{rcl_na}
\alias{rcl_na}
\title{Reclassify to NA}
\usage{
rcl_na(x, filename)
}
\arguments{
\item{x}{Input raster.}

\item{filename}{Temporary filename.}
}
\value{
Raster object.
}
\description{
Helper function that reclassifies two summed change rasters.
}
\examples{
None
}
