% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance}
\usage{
variable_importance(data, model, iterations_num = 1, clean = FALSE)
}
\arguments{
\item{data}{Input data with variables for which to calculate the variable importance. With this data you should be able to run predict function on the model.}

\item{model}{Model to be used for prediction. Function is tested only on glm object class.}

\item{iterations_num}{Number of randomization iterations. Default is 1 iteration.}

\item{clean}{Return cleaned data (default is \code{FALSE}). A dataframe will be returned, only with variables that participated in the model (in case of model selection).}
}
\value{
Output is a matrix where rows have variable importance value for each variable, and the columns are individual iterations. If clean = TRUE, return class is dataframe.
}
\description{
Calculate contribution of predictor variables to the model.
Function will make a reference prediction of the model using the standard set of variables.
Then, the values in predictor variables are randomized, and the prediction is repeated with the set of variables
that contain a randomized variable. Correlation coefficient is calculated between the reference prediction and randomized prediction.
Given importance value is \code{1 - correlation ** 2} for each variable. Number of randomizations can be set (default is one)
}
\examples{
var_importance(data = mydat, model = my_model, iterations_num = 10)
}
\author{
Mirza Cengic
}
