% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_stepwise.R
\name{vif_select_vars}
\alias{vif_select_vars}
\title{Select variables based on VIF.}
\usage{
vif_select_vars(input_rasters, points_number = 10000, thresh)
}
\arguments{
\item{input_rasters}{Raster for which to calculate VIF}

\item{points_number}{Number of sampled points}

\item{thresh}{VIF threshold.}
}
\value{
RasterStack.
}
\description{
Select variables based on VIF.
}
\examples{
None.
}
