% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_stepwise.R
\name{vif_stepwise}
\alias{vif_stepwise}
\title{VIF stepwise calculation}
\usage{
vif_stepwise(in_frame, thresh = 10, trace = FALSE, ...)
}
\arguments{
\item{in_frame}{Input data}

\item{thresh}{VIF threshold}

\item{trace}{Verbose}

\item{...}{Passed to lm function.}
}
\value{
Obj
}
\description{
VIF stepwise calculation
}
\examples{
None
}
