% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(data, aic_step = TRUE)
}
\arguments{
\item{data}{Data for modeling. Output of \code{format_data()} function.}

\item{aic_step}{Perform aic stepwise model selection. Default is \code{TRUE}.}
}
\value{
Model of "glm" class.
}
\description{
Fits model for agricultural conversion with implemented weighing scheme.
#'
}
\examples{
None.
}
