% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{get_evaluations}
\alias{get_evaluations}
\title{Get model evaluations}
\usage{
get_evaluations(fitted_model, data, ID)
}
\arguments{
\item{fitted_model}{- glm model. Output of fit_model() function (or glm() function in that matter). Any model of class "glm".}

\item{data}{- data for modeling. Output of format_data() function. List with "evaluation_data" slot.}

\item{ID}{- model_id for the current model. Stored in a variable. Region + category}
}
\value{
List with two dataframes.
}
\description{
This function wraps around \code{evaluate_model()} and returns a list that
contains \code{model_evaluation} and \code{model_coefficients} dataframes.
}
\examples{
None.

}
